[ACCESS]: 1.完善日志信息 2.防止内存泄露

+++ src/clang/lib/access/access.c
@@ -16,7 +16,7 @@ static int acc_creat_queue(acc_cntx_t *ctx);
 /* CID比较回调 */
 static int acc_conn_cid_cmp_cb(const acc_socket_extra_t *extra1, const acc_socket_extra_t *extra2)
 {
-    return (extra1->cid - extra2->cid);
+    return (int)(extra1->cid - extra2->cid);
 }
 
 /* CID哈希回调 */
